Update and rename loadImageWithSubfolders.py to loadImageWithSubfolders-NoTemp.py
Forked to add an If statement, testing whether a file begins with '._' before adding it to the file list. Added 'import re' for the regex search. 

Files that begin with '._' are temporary files that don't really exist, made by MacOS. They take up space in the file list and break things if selected by mistake.

There's probably a better way to do this, but it seems to work, and didn't cause any problems in my limited testing.